+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main / unreleased
 
+* [FEATURE] Rollout-operator can now "mirror" replicas of statefulset from any reference resource. `status.replicas` field of reference resource is kept up-to-date with current number of replicas in target statefulset. #129
+
 ## v0.13.0
 
 * [BUGFIX] Consider missing pods as not ready. #127